requirements.txt
discord.py
django-battlenet
